+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Misc:
 - Simplify `sider.yml` schema [#2192](https://github.com/sider/runners/pull/2192)
 - Secure `pip install` [#2194](https://github.com/sider/runners/pull/2194)
 - Introduce common option `target` [#2191](https://github.com/sider/runners/pull/2191)
+- **PHPMD** Handle syntax error as issue [#2201](https://github.com/sider/runners/pull/2201)
 
 ## 0.45.0